// src/types/theme.ts
export type ThemeTokens = {
  siteName?: string;
  logoPath?: string;
  faviconPath?: string;
  tocDefaultOpen?: boolean;
  contentAlign?: "justify" | "left";
  direction?: "ltr" | "rtl";
  primaryColor: string;
  linkColor: string;
  darkBackground: string;
  lightBackground: string;
  textColor: string;
  mutedTextColor: string;
  borderColor: string;
};